requirements.txt
asgiref==3.6.0
channels==4.0.0
contourpy==1.0.7
coverage==7.2.5
cycler==0.11.0
Django==4.2
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-debug-toolbar==4.0.0
djangorestframework==3.14.0
fonttools==4.39.3
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.9
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
tzdata==2023.3
